% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date_Roller.R
\name{Date_Roller}
\alias{Date_Roller}
\title{Date_Roller}
\usage{
Date_Roller(DateVec, N = 24)
}
\arguments{
\item{DateVec}{Any arranged vector of dates. Function will sort your date vector just in case you forget...}

\item{N}{Rolling N periods.}
}
\value{
Vector of dates to be used in map.
}
\description{
Creates a list of rolling date entries to be used easily using a map call.
It will start at first date, look ahead N periods and create the first list entry.
Thereafter it drops first entry, and adds next. This keeps a constant N length rolling window.
See my question on stack here: https://stackoverflow.com/q/54733427/4198868
}
\examples{
library(tidyverse);library(lubridate);library(rmsfuns)
DateVec <- rmsfuns::dateconverter(lubridate::ymd(20020101), lubridate::ymd(20200131), Transform = "weekdayEOM")
Rolldates <- Date_Roller(DateVec, N = 24)
}
