% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpts.R
\name{ggpts}
\alias{ggpts}
\title{ggpts}
\usage{
ggpts(x)
}
\arguments{
\item{x}{size}
}
\value{
changes text size in plot
}
\description{
Makes text size changes much easier in ggplot
}
\examples{
g <- ggplot(mtcars, aes(x=wt, y=mpg)) + geom_point() +
annotate(geom="label", x = 3.5, y = 22.5, label = "Some Text", size = ggpts(12))
}
