% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{make_project}
\usage{
make_project(FilePath = NULL, ProjNam = NULL, Mac = FALSE, Open = FALSE)
}
\arguments{
\item{FilePath}{Folder location on PC. Default uses clipboard (copied location)}

\item{ProjNam}{Project name Default uses last name of directory.}

\item{Mac}{If using a Mac, set to TRUE. FALSE by default}

\item{Open}{Should the project be opened or not. FALSE by Default}
}
\description{
Creates a new template project folder from a copied location. A Rproject environment with the parent folder will be created. If folder is non-empty, this will return an error (so it won't override a non-empty folder accidentally).
}
\examples{
# Create a folder, copy its location, and run:
make_proj_clipboard()
}
