% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_order_set.R
\name{plot_order_set}
\alias{plot_order_set}
\title{plot_order_set}
\usage{
plot_order_set(df, Column, Order)
}
\arguments{
\item{df}{tidy data frame to be ordered}

\item{Column}{which column to be ordered.}

\item{Order}{Vector to order by.}
}
\value{
Orders ggplot
}
\description{
Sets order for a ggplot using a column and a vector
}
\examples{
library(tidyverse);library(lubridate)
dfplot <- fmxdat::SectorTRIs \%>\% gather(type, val, -Date)
ord <- rev(unique(dfplot$type))
dfplot \%>\% ggplot() + geom_line(aes(Date, val, color = type))
# Reordered now:
dfplot \%>\% plot_order_set( Column = "type", Order = ord) \%>\% ggplot() + geom_line(aes(Date, val, color = type))
}
