% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_month_min.R
\name{safe_month_min}
\alias{safe_month_min}
\title{safe_month_min}
\usage{
safe_month_min(datesel, N = 6)
}
\arguments{
\item{Ra}{N months back}
}
\value{
N months back
}
\description{
Safely goes back N months for monthly data.
This is an improved version of using filter(date >= last(date) %-% months(6)) - c.f. below:
library(tidyverse);library(lubridate)
df <- data.frame(date = rmsfuns::dateconverter(StartDate = lubridate::ymd(20200831), EndDate = lubridate::ymd(20210228), Transform = 'calendarEOM'))
df %>% filter(date > last(date) %m-% months(6))
Note that this gives 7 months, not 6.
Instead, use:
df %>% filter(date >= safe_month_min(last(date), N = 6))
}
\examples{
library(tidyverse);library(lubridate)
df <- data.frame(date = rmsfuns::dateconverter(StartDate = lubridate::ymd(20200831), EndDate = lubridate::ymd(20210228), Transform = 'calendarEOM'))
df \%>\% filter(date >= safe_month_min(last(date), N = 6))
}
