% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_year_min.R
\name{safe_year_min}
\alias{safe_year_min}
\title{safe_year_min}
\usage{
safe_year_min(datesel, N = 2)
}
\arguments{
\item{Ra}{N months back}
}
\value{
N months back
}
\description{
Safely goes back N years for monthly data.
This is an improved version of using filter(date >= last(date) %-% months(6)) - c.f. below:
library(tidyverse);library(lubridate)
df <- data.frame(date = rmsfuns::dateconverter(StartDate = lubridate::ymd(20180831), EndDate = lubridate::ymd(20210228), Transform = 'calendarEOM'))
df %>% filter(date > last(date) %m-% years(2))
Note that this gives 25 months, not 24.
Instead, use:
df %>% filter(date >= safe_year_min(last(date), N = 2))
}
\examples{
library(tidyverse);library(lubridate)
df <- data.frame(date = rmsfuns::dateconverter(StartDate = lubridate::ymd(20180831), EndDate = lubridate::ymd(20210531), Transform = 'calendarEOM'))
df \%>\% filter(date >= safe_year_min(last(date), N = 2))
}
