% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_all.R
\name{source_all}
\alias{source_all}
\title{source_all}
\usage{
source_all(Loc)
}
\arguments{
\item{Loc}{Provided location to source all functions from
On a Windows PC, this could e.g. be: 'C:/Temp_Folder/code'
Or if called from within a project, simply source_all("code")}
}
\value{
Sources all the .R scripts in a given folder.
}
\description{
Sources all the .R scripts in a folder. Important to ensure the .R files in a folder are all functions - not running scripts.
}
\examples{
source_all("C:/Temp/FinMetrics/Practical1/code")
source_all("code")
}
