% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{make_project}
\usage{
make_project(FilePath, ProjNam)
}
\arguments{
\item{FilePath}{Provide the full location path within which the directory should be created.
On a Windows PC, this could e.g. be: 'C:/Temp_Folder/New_Project'}

\item{ProjNam}{Name of the project you want to create, e.g. 'ProjNam'}
}
\value{
Path address just built.
}
\description{
make_project builds an R folder with a workable template in a given Root directory.
}
\examples{
make_project("C:/Temp/FinMetrics/Practical1")
}
