% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fmx.R
\name{theme_fmx}
\alias{theme_fmx}
\title{theme_fmx}
\usage{
theme_fmx(
  base.size = 15,
  title.size = 14,
  subtitle.size = 12,
  caption.size = 9,
  legend.size = 10,
  axis.size.title = 10,
  axis.size = 10,
  strip.size = 10,
  legend.pos = "bottom",
  grid_Col = "#F0EFEF",
  Strip_Col = "#DDE1EE",
  CustomCaption = FALSE,
  CaptionColor = "#606060",
  FontType = "sans"
)
}
\arguments{
\item{FontType}{Check available fonts using: windowsFonts()}
}
\value{
plot theme
}
\description{
Plotting theme for ggplot: with boundary
}
\examples{
g + theme_fmx(title.size = ggpts(35), axis.size.title = ggpts(35), grid_Col = "#FFFFFF") # Remove grids
g + theme_fmx(title.size = ggpts(35), axis.size.title = ggpts(35))

# Custom Caption used as:
#   labs(caption= "First~line \n italic('and a second,')~bold('fancy one') \n 'also,'~integral(f(x)*dx, a, b)~'for good measure'")
# Bold and italics work from second line...
}
